% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_genotype_plot.R
\name{combine_genotype_plot}
\alias{combine_genotype_plot}
\title{Combine Genotype Plot Elements}
\usage{
combine_genotype_plot(genotype_plot_res, heights = c(1, 9), widths = c(2, 3))
}
\arguments{
\item{genotype_plot_res}{List of Genotype Plot elements derived from running genotype_plot function.}

\item{heights}{Set the relative heights of the snp position labels and genotypes/dendrogram respectively}

\item{widths}{Set the relative widths of the dendrogram and genotyps/snp position labels respectively}
}
\value{
A fully composed genotype plot based on all available elements
}
\description{
Function combines all present elements of the genotype_plot() output
}
